ESOLA-44 Fix firmware upgrade checking issues
We're using the bundle display version to determine a downgrade
which include the point release number causing comparison issue.
The change is to rely on WS embedded/proxy to perform the compatibility
checks so the logic is consistent.
Note: WS proxy does not allow downgrades but WS embedded does in cases,
therefore, any downgrades will need to be done using WS embedded.