ESOLA-71 Resolve irrelevant error message when segment size is different for an existing volume.
The error message seen in the issue is because the combination of the new segment size and
the volume size causes the new capacity calculation to be lower than the original capacity.
The segment size is actually not allowed to be change on existing volumes so the error should
have been thrown before the calculation and the reported error be seen.

The change is to swap the order which the checks are done so the proper error message is shown.